job implementation almost complete

+++ src/main/java/com/rbslayer/campusconnectbackend/repository/RecruiterRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RecruiterRepository extends JpaRepository<Recruiter, Integer> {
+public interface RecruiterRepository extends JpaRepository<Recruiter, Long> {
     Optional<Recruiter> findByUserId(Long userId);
 
     boolean existsByUserId(Long userId);
